requirements.txt
fake-factory==0.5.7
Flask==0.10.1
gevent==1.1.1
greenlet==0.4.9
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
locustio==0.7.3
MarkupSafe==0.23
msgpack-python==0.4.7
numpy==1.11.0
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
Werkzeug==0.11.9
wheel==0.24.0
